+++ api/src/infra/api/yuki/external.rs
@@ -71,9 +71,9 @@ pub struct YukicoderTag {
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct YukicoderStatistics {
-    pub total: i64,
+    pub total: i32,
 
-    pub solved: i64,
+    pub solved: i32,
 
     pub first_accepted_time_second: i64,
